fix Conflicts:
	LuaIOS/main.m
	LuaSDK/LuaBridgeSupport/LuaBridgeFunctor.m
	LuaSDK/LuaObjCRuntime/LuaObjCClass.m
	LuaSDK/LuaObjCRuntime/LuaObjCMessage.m

+++ LuaSDK/LuaEngine/LuaLibraryInformation.h
@@ -2,7 +2,7 @@
 //  LuaLibraryInformation.h
 //  AZenecaExhibit
 //
-//  Created by E-Reach Administrator on 6/1/12.
+//  Created by tearsofphoenix on 6/1/12.
 //  Copyright (c) 2012 __MyCompanyName__. All rights reserved.
 //
 
@@ -14,7 +14,7 @@
 
 @property (nonatomic, retain) NSString *libaName;
 @property (nonatomic) lua_CFunction loadFunction;
-@property (nonatomic) NSInteger numberOfUpvalues;
+@property (nonatomic) int numberOfUpvalues;
 @property (nonatomic, retain) NSArray *dependentLibNames;
 @property (nonatomic, retain) NSString *featureID;
 
@@ -26,13 +26,13 @@
 extern LuaLibraryInformation * LuaLibraryInformationMake(NSString *fetureID,
                                                          NSString *libName,
                                                          lua_CFunction loadFunction,
-                                                         NSInteger numberOfUpvalues,
+                                                         int numberOfUpvalues,
                                                          NSArray *dependentLibNames);
 
 extern LuaLibraryInformation * LuaLibraryInformationMakeC(NSString *fetureID,
                                                           const char *libName,
                                                           lua_CFunction loadFunction,
-                                                          NSInteger numberOfUpvalues,
+                                                          int numberOfUpvalues,
                                                           NSArray *dependentLibNames);
 
 extern void LuaLibraryInformationRegisterToState(NSDictionary *_luaEngineLibs,
